requirements.txt
# Windows 窗口控制
pywin32>=306
psutil>=5.9.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR（中文支持，GPU 加速）
paddlepaddle-gpu>=3.0.0,<3.1.0
paddleocr>=3.0.0

# 热键监听
keyboard>=0.13.5
